requirements2.txt
streamlit
torch
diffusers
transformers
accelerate
safetensors
pillow
